switch follow-btn according to conditions

+++ apps/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
+from django.views.decorators.http import require_POST
 from .models import Tweet, Relationship
 from django.contrib.auth.models import User
 from django.utils import timezone
@@ -13,24 +14,31 @@ def tweet_list(request):
 def profile(request,pk):
 	user = get_object_or_404(User, pk=pk)
 	tweets = user.tweet_set.all().order_by('-created_date')
-	following_relations = Relationship.objects.filter(from_user=user)
-	followed_relations = Relationship.objects.filter(target_user=user)
-	return render(request, 'apps/profile.html', {'user':user, 'tweets': tweets, 'following_relations': following_relations, 'followed_relations': followed_relations,})
+	follows = Relationship.objects.filter(from_user=user)
+	followers = Relationship.objects.filter(target_user=user)
+	if Relationship.objects.filter(from_user=request.user, target_user=user).exists():
+		following=True
+	else:
+		following=False
+	return render(request, 'apps/profile.html', {'user':user, 'tweets': tweets, 'follows': follows, 'followers': followers, 'following': following,})
 
 
+
+@require_POST
 def follow(request,pk):
-	from_user = request.user
-	target_user = User.objects.get(pk=pk)
-	Relationship.objects.create(from_user=from_user, target_user=target_user)
-	return redirect('apps:profile', pk=pk)
 
+	if 'follow' in request.POST:
+		from_user = request.user
+		target_user = User.objects.get(pk=pk)
+		Relationship.objects.create(from_user=from_user, target_user=target_user)
+		return redirect('apps:profile', pk=pk)
 
-def unfollow(request,pk):
-	from_user = request.user
-	target_user = User.objects.get(pk=pk)
-	relationship = Relationship.objects.filter(from_user=from_user, target_user=target_user)
-	relationship.delete()
-	return redirect('apps:profile', pk=pk)
+	elif 'unfollow' in request.POST:
+		from_user = request.user
+		target_user = User.objects.get(pk=pk)
+		relationship = Relationship.objects.filter(from_user=from_user, target_user=target_user)
+		relationship.delete()
+		return redirect('apps:profile', pk=pk)
 
 
 def tweet_new(request):
